+++ test/eca/features/prompt_test.clj
@@ -6,21 +6,25 @@
    [eca.test-helper :as h]))
 
 (deftest build-instructions-test
-  (testing "Should create instructions with rules, contexts, and behavior"
+  (testing "Should create instructions with rules, contexts, and agent behavior"
     (let [refined-contexts [{:type :file :path "foo.clj" :content "(ns foo)"}
                             {:type :file :path "bar.clj" :content "(def a 1)" :lines-range {:start 1 :end 1}}
                             {:type :repoMap}
                             {:type :mcpResource :uri "custom://my-resource" :content "some-cool-content"}]
           rules [{:name "rule1" :content "First rule"}
                  {:name "rule2" :content "Second rule"}]
+          skills [{:name "review-pr" :description "Review a PR"}
+                  {:name "lint-fix" :description "Fix a lint"}]
           fake-repo-map (delay "TREE")
           behavior "agent"
           config {}
-          result (prompt/build-chat-instructions refined-contexts rules fake-repo-map behavior config nil [] (h/db))]
+          result (prompt/build-chat-instructions refined-contexts rules skills fake-repo-map behavior config nil [] (h/db))]
       (is (string/includes? result "You are ECA"))
       (is (string/includes? result "<rules description=\"Rules defined by user\">"))
       (is (string/includes? result "<rule name=\"rule1\">First rule</rule>"))
       (is (string/includes? result "<rule name=\"rule2\">Second rule</rule>"))
+      (is (string/includes? result "<skill name=\"review-pr\" description=\"Review a PR\"/>"))
+      (is (string/includes? result "<skill name=\"lint-fix\" description=\"Fix a lint\"/>"))
       (is (string/includes? result "<contexts description=\"User-Provided. This content is current and accurate. Treat this as sufficient context for answering the query.\">"))
       (is (string/includes? result "<file path=\"foo.clj\">(ns foo)</file>"))
       (is (string/includes? result "<file line-start=1 line-end=1 path=\"bar.clj\">(def a 1)</file>"))
@@ -35,14 +39,18 @@
                             {:type :mcpResource :uri "custom://my-resource" :content "some-cool-content"}]
           rules [{:name "rule1" :content "First rule"}
                  {:name "rule2" :content "Second rule"}]
+          skills [{:name "review-pr" :description "Review a PR"}
+                  {:name "lint-fix" :description "Fix a lint"}]
           fake-repo-map (delay "TREE")
           behavior "plan"
           config {}
-          result (prompt/build-chat-instructions refined-contexts rules fake-repo-map behavior config nil [] (h/db))]
+          result (prompt/build-chat-instructions refined-contexts rules skills fake-repo-map behavior config nil [] (h/db))]
       (is (string/includes? result "You are ECA"))
       (is (string/includes? result "<rules description=\"Rules defined by user\">"))
       (is (string/includes? result "<rule name=\"rule1\">First rule</rule>"))
       (is (string/includes? result "<rule name=\"rule2\">Second rule</rule>"))
+      (is (string/includes? result "<skill name=\"review-pr\" description=\"Review a PR\"/>"))
+      (is (string/includes? result "<skill name=\"lint-fix\" description=\"Fix a lint\"/>"))
       (is (string/includes? result "<contexts description=\"User-Provided. This content is current and accurate. Treat this as sufficient context for answering the query.\">"))
       (is (string/includes? result "<file path=\"foo.clj\">(ns foo)</file>"))
       (is (string/includes? result "<file line-start=1 line-end=1 path=\"bar.clj\">(def a 1)</file>"))